Split Stress Test Orchestrator and Runner
