+++ packages/test/service-load-test/package.json
@@ -53,6 +53,7 @@
   },
   "dependencies": {
     "@fluidframework/aqueduct": "^0.37.0",
+    "@fluidframework/common-utils": "^0.28.0-0",
     "@fluidframework/container-definitions": "^0.37.0",
     "@fluidframework/container-loader": "^0.37.0",
     "@fluidframework/core-interfaces": "^0.37.0",
