+++ packages/test/service-load-test/src/nodeStressTest.ts
@@ -3,88 +3,18 @@
  * Licensed under the MIT License.
  */
 
-import fs from "fs";
 import child_process from "child_process";
-import assert from "assert";
 import commander from "commander";
-import { Loader } from "@fluidframework/container-loader";
-import { IFluidCodeDetails } from "@fluidframework/core-interfaces";
-import { LocalCodeLoader } from "@fluidframework/test-utils";
-import { ITestDriver, TestDriverTypes, ITelemetryBufferedLogger } from "@fluidframework/test-driver-definitions";
-import { createFluidTestDriver } from "@fluidframework/test-drivers";
-import { requestFluidObject } from "@fluidframework/runtime-utils";
-import { pkgName, pkgVersion } from "./packageVersion";
-import { ITestConfig, ILoadTestConfig } from "./testConfigFile";
-import { IRunConfig, fluidExport, ILoadTest } from "./loadTestDataStore";
-
-const packageName = `${pkgName}@${pkgVersion}`;
-
-// Provide default implementation of getTestLogger since it's declared as always defined
-const nullLogger: ITelemetryBufferedLogger = { send: () => {}, flush: async () => {} };
-(global as any).getTestLogger = () => nullLogger;
-
-let logger: ITelemetryBufferedLogger = getTestLogger();
-
-const codeDetails: IFluidCodeDetails = {
-    package: packageName,
-    config: {},
-};
-
-const codeLoader = new LocalCodeLoader([[codeDetails, fluidExport]]);
-
-function createLoader(testDriver: ITestDriver) {
-    // Construct the loader
-    const loader = new Loader({
-        urlResolver: testDriver.createUrlResolver(),
-        documentServiceFactory: testDriver.createDocumentServiceFactory(),
-        codeLoader,
-        logger,
-    });
-    return loader;
-}
-
-async function initialize(testDriver: ITestDriver) {
-    const loader = createLoader(testDriver);
-    const container = await loader.createDetachedContainer(codeDetails);
-    container.on("error", (error) => {
-        console.log(error);
-        process.exit(-1);
-    });
-    const testId = Date.now().toString();
-    const request = testDriver.createCreateNewRequest(testId);
-    await container.attach(request);
-    container.close();
-
-    return testId;
-}
-
-async function load(testDriver: ITestDriver, testId: string) {
-    const loader = createLoader(testDriver);
-    const url =  await testDriver.createContainerUrl(testId);
-    const container = await loader.resolve({ url });
-    return requestFluidObject<ILoadTest>(container,"/");
-}
-
-const createTestDriver =
-    async (driver: TestDriverTypes) => createFluidTestDriver(driver,{
-        odsp: {
-            directory: "stress",
-        },
-    });
+import { TestDriverTypes } from "@fluidframework/test-driver-definitions";
+import { ILoadTestConfig } from "./testConfigFile";
+import { createTestDriver, getProfile, initialize, safeExit } from "./utils";
 
 async function main() {
-    if (process.env.FLUID_TEST_LOGGER_PKG_PATH) {
-        await import(process.env.FLUID_TEST_LOGGER_PKG_PATH);
-        logger = getTestLogger();
-        assert(logger, "Expected getTestLogger to return something");
-    }
-
     commander
         .version("0.0.1")
         .requiredOption("-d, --driver <driver>", "Which test driver info to use", "odsp")
         .requiredOption("-p, --profile <profile>", "Which test profile to use from testConfig.json", "ci")
         .option("-id, --testId <testId>", "Load an existing data store rather than creating new")
-        .option("-r, --runId <runId>", "run a child process with the given id. Requires --testId option.")
         .option("-dbg, --debug", "Debug child processes via --inspect-brk")
         .option("-l, --log <filter>", "Filter debug logging. If not provided, uses DEBUG env variable.")
         .parse(process.argv);
@@ -92,82 +22,22 @@ async function main() {
     const driver: TestDriverTypes = commander.driver;
     const profileArg: string = commander.profile;
     const testId: string | undefined = commander.testId;
-    const runId: number | undefined = commander.runId === undefined ? undefined : parseInt(commander.runId, 10);
     const debug: true | undefined = commander.debug;
     const log: string | undefined = commander.log;
 
-    let config: ITestConfig;
-    try {
-        config = JSON.parse(fs.readFileSync("./testConfig.json", "utf-8"));
-    } catch (e) {
-        console.error("Failed to read testConfig.json");
-        console.error(e);
-        process.exit(-1);
-    }
-
-    const profile: ILoadTestConfig | undefined = config.profiles[profileArg];
-    if (profile === undefined) {
-        console.error("Invalid --profile argument not found in testConfig.json profiles");
-        process.exit(-1);
-    }
+    const profile = getProfile(profileArg);
 
     if (log !== undefined) {
         process.env.DEBUG = log;
     }
 
-    let result: number;
-    // When runId is specified (with testId), kick off a single test runner and exit when it's finished
-    if (runId !== undefined) {
-        if (testId === undefined) {
-            console.error("Missing --testId argument needed to run child process");
-            process.exit(-1);
-        }
-        result = await runnerProcess(driver, profile, runId, testId);
-    }
-    else {
-        // When runId is not specified, this is the orchestrator process which will spawn child test runners.
-        result = await orchestratorProcess(
-            driver,
-            { ...profile, name: profileArg },
-            { testId, debug });
-    }
+    const result = await orchestratorProcess(
+        driver,
+        { ...profile, name: profileArg },
+        { testId, debug });
 
-    // There seems to be at least one dangling promise in ODSP Driver, give it a second to resolve
-    await(new Promise((res) => { setTimeout(res, 1000); }));
-    // Flush the logs
-    await logger.flush();
-
-    process.exit(result);
+    await safeExit(result);
 }
-
-/**
- * Implementation of the runner process. Returns the return code to exit the process with.
- */
-async function runnerProcess(
-    driver: TestDriverTypes,
-    profile: ILoadTestConfig,
-    runId: number,
-    testId: string,
-): Promise<number> {
-    try {
-        const runConfig: IRunConfig = {
-            runId,
-            testConfig: profile,
-        };
-
-        const testDriver = await createTestDriver(driver);
-
-        const stressTest = await load(testDriver, testId);
-        await stressTest.run(runConfig);
-        console.log(`${runId.toString().padStart(3)}> exit`);
-        return 0;
-    } catch (e) {
-        console.error(`${runId.toString().padStart(3)}> error: loading test`);
-        console.error(e);
-        return -1;
-    }
-}
-
 /**
  * Implementation of the orchestrator process. Returns the return code to exit the process with.
  */
@@ -189,7 +59,7 @@ async function orchestratorProcess(
     const p: Promise<void>[] = [];
     for (let i = 0; i < profile.numClients; i++) {
         const childArgs: string[] = [
-            "./dist/nodeStressTest.js",
+            "./dist/runner.js",
             "--driver", driver,
             "--profile", profile.name,
             "--runId", i.toString(),
